Update and rename ClientsSukodu2.py to SudokuClient2.py
